refactor single call class name
