+++ chat/kit/src/main/java/cn/wildfire/chat/kit/voip/SingleCallActivity.java
@@ -10,44 +10,26 @@
 
 package cn.wildfire.chat.kit.voip;
 
-import android.annotation.TargetApi;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.PowerManager;
-import android.provider.Settings;
 import android.util.Log;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowManager.LayoutParams;
-import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
 import org.webrtc.StatsReport;
 
 import java.util.Collections;
-import java.util.List;
 
-import butterknife.ButterKnife;
 import cn.wildfirechat.avenginekit.AVEngineKit;
-import cn.wildfirechat.chat.R;
-import cn.wildfirechat.client.NotInitializedExecption;
 import cn.wildfirechat.model.Conversation;
 
 /**
  * Activity for peer connection call setup, call waiting
  * and call view.
  */
-public class SingleVoipCallActivity extends FragmentActivity implements AVEngineKit.CallSessionCallback {
+public class SingleCallActivity extends VoipBaseActivity {
     private static final String TAG = "P2PVideoActivity";
     private static final int REQUEST_CODE_DRAW_OVERLAY = 100;
 
@@ -56,84 +38,26 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
     public static final String EXTRA_AUDIO_ONLY = "audioOnly";
     public static final String EXTRA_FROM_FLOATING_VIEW = "fromFloatingView";
 
-    // List of mandatory application permissions.
-    private static final String[] MANDATORY_PERMISSIONS = {"android.permission.MODIFY_AUDIO_SETTINGS",
-            "android.permission.RECORD_AUDIO", "android.permission.INTERNET"};
-
-    private AVEngineKit gEngineKit;
-
-    protected PowerManager.WakeLock wakeLock;
-
-
     private boolean isOutgoing;
     private String targetId;
     private boolean isAudioOnly;
     private boolean isFromFloatingView;
-    private Handler handler = new Handler();
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-//    Thread.setDefaultUncaughtExceptionHandler(new UnhandledExceptionHandler(this));
-
-        // Set window styles for fullscreen-window size. Needs to be done before
-        // adding content.
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
-        PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
-        wakeLock = powerManager.newWakeLock(
-                PowerManager.ACQUIRE_CAUSES_WAKEUP |
-                        PowerManager.SCREEN_BRIGHT_WAKE_LOCK, "bright");
-        if (wakeLock != null) {
-            wakeLock.acquire();
-        }
-
-        getWindow().addFlags(LayoutParams.FLAG_FULLSCREEN | LayoutParams.FLAG_KEEP_SCREEN_ON
-                | LayoutParams.FLAG_SHOW_WHEN_LOCKED | LayoutParams.FLAG_TURN_SCREEN_ON);
-        getWindow().getDecorView().setSystemUiVisibility(getSystemUiVisibility());
-        setContentView(R.layout.av_p2p_video_activity);
-
-        try {
-            gEngineKit = AVEngineKit.Instance();
-        } catch (NotInitializedExecption notInitializedExecption) {
-            notInitializedExecption.printStackTrace();
-            finish();
-        }
-        ButterKnife.bind(this);
-
-        // Check for mandatory permissions.
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            for (String permission : MANDATORY_PERMISSIONS) {
-                if (checkCallingOrSelfPermission(permission) != PackageManager.PERMISSION_GRANTED) {
-                    requestPermissions(MANDATORY_PERMISSIONS, 100);
-                    break;
-                }
-            }
-        }
-
-        // TODO init fragment
         final Intent intent = getIntent();
 
         targetId = intent.getStringExtra(EXTRA_TARGET);
         isFromFloatingView = intent.getBooleanExtra(EXTRA_FROM_FLOATING_VIEW, false);
         if (isFromFloatingView) {
-            Intent serviceIntent = new Intent(this, FloatingSingleCallService.class);
+            Intent serviceIntent = new Intent(this, SingleCallFloatingService.class);
             stopService(serviceIntent);
             initFromFloatView();
         } else {
             isOutgoing = intent.getBooleanExtra(EXTRA_MO, false);
             isAudioOnly = intent.getBooleanExtra(EXTRA_AUDIO_ONLY, false);
             init(targetId, isOutgoing, isAudioOnly);
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        for (int result : grantResults) {
-            if (result != PackageManager.PERMISSION_GRANTED) {
-                Toast.makeText(this, "需要录音和摄像头权限，才能进行语音通话", Toast.LENGTH_SHORT).show();
-                finish();
-                return;
-            }
         }
     }
 
@@ -145,14 +69,14 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
             finish();
             return;
         } else {
-            session.setCallback(SingleVoipCallActivity.this);
+            session.setCallback(SingleCallActivity.this);
         }
 
         Fragment fragment;
         if (session.isAudioOnly()) {
-            fragment = new AudioFragment();
+            fragment = new SingleAudioFragment();
         } else {
-            fragment = new VideoFragment();
+            fragment = new SingleVideoFragment();
         }
 
         currentCallback = (AVEngineKit.CallSessionCallback) fragment;
@@ -167,9 +91,9 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
 
         Fragment fragment;
         if (audioOnly) {
-            fragment = new AudioFragment();
+            fragment = new SingleAudioFragment();
         } else {
-            fragment = new VideoFragment();
+            fragment = new SingleVideoFragment();
         }
 
         currentCallback = (AVEngineKit.CallSessionCallback) fragment;
@@ -181,7 +105,7 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
         if (outgoing) {
             try {
                 Conversation conversation = new Conversation(Conversation.ConversationType.Single, targetId);
-                gEngineKit.startCall(conversation, Collections.singletonList(targetId), audioOnly, SingleVoipCallActivity.this);
+                gEngineKit.startCall(conversation, Collections.singletonList(targetId), audioOnly, SingleCallActivity.this);
                 session.startPreview();
             } catch (Exception e) {
                 e.printStackTrace();
@@ -191,18 +115,9 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
             if (session == null) {
                 finish();
             } else {
-                session.setCallback(SingleVoipCallActivity.this);
+                session.setCallback(SingleCallActivity.this);
             }
         }
-    }
-
-    @TargetApi(19)
-    private static int getSystemUiVisibility() {
-        int flags = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION | View.SYSTEM_UI_FLAG_FULLSCREEN;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            flags |= View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY;
-        }
-        return flags;
     }
 
     // Activity interfaces
@@ -211,7 +126,7 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
         super.onStop();
         AVEngineKit.CallSession session = gEngineKit.getCurrentSession();
         if (session != null && session.getState() != AVEngineKit.CallState.Idle) {
-//      session.stopVideoSource();
+//            session.stopVideoSource();
         }
     }
 
@@ -226,25 +141,16 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
 
     @Override
     protected void onDestroy() {
-        Thread.setDefaultUncaughtExceptionHandler(null);
         // TODO do not endCall
 //        AVEngineKit.CallSession session = gEngineKit.getCurrentSession();
 //        if (session != null && session.getState() != AVEngineKit.CallState.Idle) {
 //            session.endCall();
 //        }
         super.onDestroy();
-        if (wakeLock != null) {
-            wakeLock.release();
-        }
     }
 
     public AVEngineKit getEngineKit() {
         return gEngineKit;
-    }
-
-    @Override
-    public void didCallEndWithReason(AVEngineKit.CallEndReason reason) {
-        finish();
     }
 
     @Override
@@ -262,7 +168,7 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
         postAction(() -> {
             currentCallback.didChangeMode(audioOnly);
             if (audioOnly) {
-                AudioFragment fragment = new AudioFragment();
+                SingleAudioFragment fragment = new SingleAudioFragment();
                 FragmentManager fragmentManager = getSupportFragmentManager();
                 fragmentManager.beginTransaction()
                         .replace(android.R.id.content, fragment)
@@ -282,16 +188,6 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
     }
 
     @Override
-    public void didParticipantJoined(String s) {
-
-    }
-
-    @Override
-    public void didParticipantLeft(String s, AVEngineKit.CallEndReason callEndReason) {
-
-    }
-
-    @Override
     public void didCreateLocalVideoTrack() {
         postAction(() -> currentCallback.didCreateLocalVideoTrack());
     }
@@ -301,13 +197,8 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
         postAction(() -> currentCallback.didReceiveRemoteVideoTrack(s));
     }
 
-    @Override
-    public void didRemoveRemoteVideoTrack(String s) {
-
-    }
-
     public void audioAccept() {
-        AudioFragment fragment = new AudioFragment();
+        SingleAudioFragment fragment = new SingleAudioFragment();
         FragmentManager fragmentManager = getSupportFragmentManager();
         fragmentManager.beginTransaction()
                 .replace(android.R.id.content, fragment)
@@ -345,32 +236,11 @@ public class SingleVoipCallActivity extends FragmentActivity implements AVEngine
             return;
         }
 
-        Intent intent = new Intent(this, FloatingSingleCallService.class);
+        Intent intent = new Intent(this, SingleCallFloatingService.class);
         intent.putExtra(EXTRA_TARGET, targetId);
         intent.putExtra(EXTRA_AUDIO_ONLY, isAudioOnly);
         intent.putExtra(EXTRA_MO, isOutgoing);
         startService(intent);
         finish();
-    }
-
-    private void postAction(Runnable action) {
-        handler.post(action);
-    }
-
-    private boolean checkOverlayPermission() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (!Settings.canDrawOverlays(this)) {
-                Toast.makeText(this, "需要悬浮窗权限", Toast.LENGTH_LONG).show();
-                Intent intent = new Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION, Uri.parse("package:" + getPackageName()));
-
-                List<ResolveInfo> infos = getPackageManager().queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY);
-                if (infos == null || infos.isEmpty()) {
-                    return true;
-                }
-                startActivity(intent);
-                return false;
-            }
-        }
-        return true;
     }
 }
