+++ chat/kit/src/main/java/cn/wildfire/chat/kit/voip/SingleCallFloatingService.java
@@ -34,7 +34,7 @@ import cn.wildfirechat.chat.R;
 
 import static org.webrtc.RendererCommon.ScalingType.SCALE_ASPECT_BALANCED;
 
-public class FloatingSingleCallService extends Service {
+public class SingleCallFloatingService extends Service {
     private static boolean isStarted = false;
     private static final int NOTIFICATION_ID = 1;
 
@@ -70,12 +70,12 @@ public class FloatingSingleCallService extends Service {
             stopSelf();
         }
 
-        resumeActivityIntent = new Intent(this, SingleVoipCallActivity.class);
-        resumeActivityIntent.putExtra(SingleVoipCallActivity.EXTRA_FROM_FLOATING_VIEW, true);
-        resumeActivityIntent.putExtra(SingleVoipCallActivity.EXTRA_MO, intent.getBooleanExtra(SingleVoipCallActivity.EXTRA_MO, false));
-        resumeActivityIntent.putExtra(SingleVoipCallActivity.EXTRA_AUDIO_ONLY, intent.getBooleanExtra(SingleVoipCallActivity.EXTRA_AUDIO_ONLY, false));
-        targetId = intent.getStringExtra(SingleVoipCallActivity.EXTRA_TARGET);
-        resumeActivityIntent.putExtra(SingleVoipCallActivity.EXTRA_TARGET, targetId);
+        resumeActivityIntent = new Intent(this, SingleCallActivity.class);
+        resumeActivityIntent.putExtra(SingleCallActivity.EXTRA_FROM_FLOATING_VIEW, true);
+        resumeActivityIntent.putExtra(SingleCallActivity.EXTRA_MO, intent.getBooleanExtra(SingleCallActivity.EXTRA_MO, false));
+        resumeActivityIntent.putExtra(SingleCallActivity.EXTRA_AUDIO_ONLY, intent.getBooleanExtra(SingleCallActivity.EXTRA_AUDIO_ONLY, false));
+        targetId = intent.getStringExtra(SingleCallActivity.EXTRA_TARGET);
+        resumeActivityIntent.putExtra(SingleCallActivity.EXTRA_TARGET, targetId);
         resumeActivityIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
         PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, resumeActivityIntent, PendingIntent.FLAG_UPDATE_CURRENT);
