+++ chat/kit/src/main/java/cn/wildfire/chat/kit/voip/VoipBaseActivity.java
@@ -1,0 +1,188 @@
+/*
+ *  Copyright 2015 The WebRTC Project Authors. All rights reserved.
+ *
+ *  Use of this source code is governed by a BSD-style license
+ *  that can be found in the LICENSE file in the root of the source
+ *  tree. An additional intellectual property rights grant can be found
+ *  in the file PATENTS.  All contributing project authors may
+ *  be found in the AUTHORS file in the root of the source tree.
+ */
+
+package cn.wildfire.chat.kit.voip;
+
+import android.annotation.TargetApi;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.content.pm.ResolveInfo;
+import android.net.Uri;
+import android.os.Build;
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.PowerManager;
+import android.provider.Settings;
+import android.view.View;
+import android.view.Window;
+import android.view.WindowManager.LayoutParams;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.FragmentActivity;
+
+import org.webrtc.StatsReport;
+
+import java.util.List;
+
+import cn.wildfirechat.avenginekit.AVEngineKit;
+import cn.wildfirechat.client.NotInitializedExecption;
+
+/**
+ * Activity for peer connection call setup, call waiting
+ * and call view.
+ */
+public class VoipBaseActivity extends FragmentActivity implements AVEngineKit.CallSessionCallback {
+
+    // List of mandatory application permissions.
+    private static final String[] MANDATORY_PERMISSIONS = {"android.permission.MODIFY_AUDIO_SETTINGS",
+            "android.permission.RECORD_AUDIO", "android.permission.INTERNET"};
+
+    protected AVEngineKit gEngineKit;
+    protected PowerManager.WakeLock wakeLock;
+    private Handler handler = new Handler();
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+//    Thread.setDefaultUncaughtExceptionHandler(new UnhandledExceptionHandler(this));
+
+        // Set window styles for fullscreen-window size. Needs to be done before
+        // adding content.
+        requestWindowFeature(Window.FEATURE_NO_TITLE);
+        PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
+        wakeLock = powerManager.newWakeLock(
+                PowerManager.ACQUIRE_CAUSES_WAKEUP |
+                        PowerManager.SCREEN_BRIGHT_WAKE_LOCK, "bright");
+        if (wakeLock != null) {
+            wakeLock.acquire();
+        }
+
+        getWindow().addFlags(LayoutParams.FLAG_FULLSCREEN | LayoutParams.FLAG_KEEP_SCREEN_ON
+                | LayoutParams.FLAG_SHOW_WHEN_LOCKED | LayoutParams.FLAG_TURN_SCREEN_ON);
+        getWindow().getDecorView().setSystemUiVisibility(getSystemUiVisibility());
+
+        try {
+            gEngineKit = AVEngineKit.Instance();
+        } catch (NotInitializedExecption notInitializedExecption) {
+            notInitializedExecption.printStackTrace();
+            finish();
+        }
+
+        // Check for mandatory permissions.
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            for (String permission : MANDATORY_PERMISSIONS) {
+                if (checkCallingOrSelfPermission(permission) != PackageManager.PERMISSION_GRANTED) {
+                    requestPermissions(MANDATORY_PERMISSIONS, 100);
+                    break;
+                }
+            }
+        }
+
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        for (int result : grantResults) {
+            if (result != PackageManager.PERMISSION_GRANTED) {
+                Toast.makeText(this, "需要录音和摄像头权限，才能进行语音通话", Toast.LENGTH_SHORT).show();
+                finish();
+                return;
+            }
+        }
+    }
+
+    @TargetApi(19)
+    private static int getSystemUiVisibility() {
+        int flags = View.SYSTEM_UI_FLAG_HIDE_NAVIGATION | View.SYSTEM_UI_FLAG_FULLSCREEN;
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+            flags |= View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY;
+        }
+        return flags;
+    }
+
+    @Override
+    protected void onDestroy() {
+        Thread.setDefaultUncaughtExceptionHandler(null);
+        super.onDestroy();
+        if (wakeLock != null) {
+            wakeLock.release();
+        }
+    }
+
+    public AVEngineKit getEngineKit() {
+        return gEngineKit;
+    }
+
+    @Override
+    public void didCallEndWithReason(AVEngineKit.CallEndReason reason) {
+        finish();
+    }
+
+    @Override
+    public void didError(String error) {
+    }
+
+    @Override
+    public void didGetStats(StatsReport[] reports) {
+    }
+
+    @Override
+    public void didChangeMode(boolean audioOnly) {
+    }
+
+    @Override
+    public void didChangeState(AVEngineKit.CallState state) {
+    }
+
+    @Override
+    public void didParticipantJoined(String s) {
+
+    }
+
+    @Override
+    public void didParticipantLeft(String s, AVEngineKit.CallEndReason callEndReason) {
+
+    }
+
+    @Override
+    public void didCreateLocalVideoTrack() {
+    }
+
+    @Override
+    public void didReceiveRemoteVideoTrack(String s) {
+    }
+
+    @Override
+    public void didRemoveRemoteVideoTrack(String s) {
+
+    }
+
+    protected void postAction(Runnable action) {
+        handler.post(action);
+    }
+
+    protected boolean checkOverlayPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            if (!Settings.canDrawOverlays(this)) {
+                Toast.makeText(this, "需要悬浮窗权限", Toast.LENGTH_LONG).show();
+                Intent intent = new Intent(Settings.ACTION_MANAGE_OVERLAY_PERMISSION, Uri.parse("package:" + getPackageName()));
+
+                List<ResolveInfo> infos = getPackageManager().queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY);
+                if (infos == null || infos.isEmpty()) {
+                    return true;
+                }
+                startActivity(intent);
+                return false;
+            }
+        }
+        return true;
+    }
+}
